minimize-defconfig: add new configuration to remove stuff from distro
Rather than base the .config on 'make defconfig', this fragment will
remove stuff that's not really necessary in a distro.

This helps in identifying dependencies and what is required to
boot/startup in some distributions.

Usage:
add .config in build directory
<path>/kconfig-minimize-distro-add-sof-defaults.sh

Tested with Fedora34 default configuration w/ SOF defaults added to
run on TGL-SoundWire devices.

FIXME: the memory config seems really critical to boot Fedora, not
sure the options can be used for other distros.

Signed-off-by: Pierre-Louis Bossart <pierre-louis.bossart@linux.intel.com>