base-config: add FRAMEBUFFER_CONSOLE
When graphics fail, a blank screen is not useful.

Even when graphics work, without a framebuffer console chvt still
works but it is "blind" / with a frozen GUI instead! Incredibly
confusing. Linux distributions expect this.

Also useful in case of hangs like
https://github.com/thesofproject/linux/issues/3387: text consoles
are always more responsive and they may even have some logs

Signed-off-by: Marc Herbert <marc.herbert@intel.com>